+++ google/cloud/cloudbuild/v1/cloud_build_connection_idempotency_policy.h
@@ -93,6 +93,10 @@ class CloudBuildConnectionIdempotencyPolicy {
 
   virtual google::cloud::Idempotency ListWorkerPools(
       google::devtools::cloudbuild::v1::ListWorkerPoolsRequest request);
+
+  virtual google::cloud::Idempotency GetDefaultServiceAccount(
+      google::devtools::cloudbuild::v1::GetDefaultServiceAccountRequest const&
+          request);
 };
 
 std::unique_ptr<CloudBuildConnectionIdempotencyPolicy>